+++ tests/Integration/TestHelpers.php
@@ -69,6 +69,11 @@ class TestHelpers
         return getenv('LAPI_URL');
     }
 
+    public static function getAppSecUrl(): string
+    {
+        return getenv('APPSEC_URL');
+    }
+
     public static function getBouncerKey(): string
     {
         if ($bouncerKey = getenv('BOUNCER_KEY')) {